Move WPT script to test/wpt-node
